restructure routes package
