+++ api/routes/__init__.py
@@ -1,8 +1,1 @@
 from .router import router
-from .routes import (
-    POST_ROUTES,
-    COMMENT_ROUTES,
-    VOTE_ROUTES,
-    ATTACHMENT_ROUTES,
-    UTILITY_ROUTES
-)
