--- api/routes/handlers/__init__.py
@@ -1,1 +1,0 @@
-# This module contains the handler functions for the API routes.
