+++ api/routes/votes/handlers.py
@@ -1,30 +1,29 @@
 from fastapi import Depends
-from models.clients import AWSClients, get_aws_clients
+from services.aws_clients import AWSClients, get_aws_clients
 
 # =========================
-# |  ATTACHMENT HANDLERS  |
+# |     VOTE HANDLERS     |
 # =========================
-async def upload_file(
+async def vote_post(
     post_id: int,
     aws_clients: AWSClients = Depends(get_aws_clients)
 ):
     pass
 
-async def get_post_files(
+async def remove_post_vote(
     post_id: int,
     aws_clients: AWSClients = Depends(get_aws_clients)
 ):
     pass
 
-async def delete_file(
-    post_id: int,
-    file_id: int,
+async def vote_comment(
+    comment_id: int,
     aws_clients: AWSClients = Depends(get_aws_clients)
 ):
     pass
 
-async def get_file_meta(
-    file_id: int,
+async def remove_comment_vote(
+    comment_id: int,
     aws_clients: AWSClients = Depends(get_aws_clients)
 ):
     pass
